NativeBase installation + NavBar creation & Page skeletons
Installed NativeBase dependencies. Created the NativeBaseStyling file which contains a few options for styling Text, and Button wrapper classes from NativeBase.
Created the Navigation Bar component, which is incomplete as we do not yet have Router setup.
Created the Login and Routes pages,  Login is complete except for the API call, and Routes is currently a skeletal form