Compress images and style components
Added a new function in `utils` called CompressImage that does exactly what the name suggests. This seems to run pretty quickly but for edge cases I added in a new useState that will activate when the compression function starts and will remove the select image button from the screen.

Added a max height of 90vh to the CreateRoute parent popup, replaced the base input html component with a style to mimic our native base buttons.